--- components/Modals/UserActionsModal.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import { View, Text, TouchableOpacity, Modal, ScrollView } from 'react-native';
-import { generalStyles, colors } from '../../App.styles';
-import AccountStyles from '../Account/Account.styles';
-
-export const UserActionsModal = ({
-
-}) => (
-  
-)
